% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime.R
\name{get_datetime}
\alias{get_datetime}
\title{Get Datetime}
\usage{
get_datetime(raw_unix_times, timezone = "Australia/Darwin")
}
\arguments{
\item{raw_unix_times:}{A list of unix times of messages}

\item{timezone:}{The timezone of the messages, default to Adelaide}
}
\value{
A list of datetime objects
}
\description{
Converts a list of raw unix times to datetime objects in the given timezone.
}
