% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_person_message_times.R
\name{get_person_message_times}
\alias{get_person_message_times}
\title{Get Message Times by Person}
\usage{
get_person_message_times(msgs, name)
}
\arguments{
\item{msgs:}{A list of messages of multiple users to be filtered.}

\item{name:}{The full name of the person to be extracted.}
}
\value{
A list of message times in unix timecode format.
}
\description{
Finds all of the message times for a given person in a series of messages.
}
\examples{
: get_person_message_times(msgs, "Tobin South")

}
