% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_by_person.R
\name{get_time_by_person}
\alias{get_time_by_person}
\title{Get Time by Person}
\usage{
get_time_by_person(row, name)
}
\arguments{
\item{row:}{The message meta data from a messenger conversation.}

\item{name:}{The name of the person to be checked.}
}
\value{
The unix time of the message. Or NA if it is not the correct person.
}
\description{
Finds the time a message was sent, if that user was the desired user.
}
