% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_facebook_message_data.R
\name{load_facebook_message_data}
\alias{load_facebook_message_data}
\title{Load Facebook Message Data}
\usage{
load_facebook_message_data(file_path)
}
\arguments{
\item{file_path:}{The path to the JSON file of messages.}
}
\value{
A list of all messages.
}
\description{
Loads in the data file for a chat on facebook messenger.
This data should be in JSON format and can be retrieved
from Facebook.
}
\details{
https://www.facebook.com/help/1701730696756992?helpref=hc_global_nav
}
