% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.over_time.R
\name{plot.over_time}
\alias{plot.over_time}
\title{Plot of messages over time}
\usage{
\method{plot}{over_time}(raw_datetime)
}
\arguments{
\item{raw_datetime:}{A list of datetime objects of messages.}
}
\description{
Plots the distribution of messages over time.
}
