% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.time_of_day.R
\name{plot.time_of_day}
\alias{plot.time_of_day}
\title{Plot of messages over time}
\usage{
\method{plot}{time_of_day}(raw_datetime)
}
\arguments{
\item{raw_datetime:}{A list of datetime objects of messages.}
}
\description{
Plots the distribution of messages over the course of a
normal day in the timezone of the datetime objects.
}
