% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_over_time.R
\name{plot_over_time}
\alias{plot_over_time}
\title{Plot of messages over time}
\usage{
plot_over_time(raw_datetime)
}
\arguments{
\item{raw_datetime:}{A list of datetime objects of messages.}
}
\description{
Plots the distribution of messages over time.
}
